+++ src/graph_color.h
@@ -6,7 +6,7 @@
 #ifndef MAUL2D_SRC_GRAPH_COLOR_H
 #define MAUL2D_SRC_GRAPH_COLOR_H
 
-#include "solver.h"
+#include "contact_solver.h"
 
 void m2ColorConstraints(m2World* world, m2ContactConstraint* constraints, int32_t count,
                         int32_t* colorStart);